+++ src/feature/post/index.ts
@@ -1,4 +1,5 @@
 export { getMyPosts, useGetMyPostsQuery } from './api/post-api'
+export { useDeletePostMutation } from './api/post-api'
 export {
   getAllPublicPosts,
   getPublicPost,
@@ -7,5 +8,10 @@ export {
   useGetPublicPostQuery,
   useGetUsersPostsQuery,
 } from './api/public-posts-api'
-export { CreatePostDialog } from './ui/CreatePostDialog/CreatePostDialog'
-export { PostDetailsDialogs } from './ui/PostDetailsDialogs/PostDetailsDialogs'
+export { useEditPost } from './model/hooks/useEditPost'
+export type { CurrentWindow } from './model/types/post.types'
+export { CropperPostScreen } from './ui/CropperPostScreen/CropperPostScreen'
+export { DescriptionScreen } from './ui/DescriptionScreen/DescriptionScreen'
+export { EditPostForm } from './ui/EditPostForm/EditPostForm'
+export { FiltersScreen } from './ui/FiltersScreen/FiltersScreen'
+export { PostDetails } from './ui/PostDetails/PostDetails'
