+++ src/feature/profile/index.ts
@@ -5,4 +5,8 @@ export {
   useGetPublicProfileQuery,
   useGetTotalUsersCountQuery,
 } from './api/public-profile-api'
-export { GeneralInformation } from './ui/GeneralInformation/GeneralInformation'
+export { useChangePhotoProfile } from './model/hooks/useChangePhotoProfile'
+export { useRemovePhotoProfile } from './model/hooks/useRemovePhotoProfile'
+export { useUpdateProfile } from './model/hooks/useUpdateProfile'
+export { EditProfileForm } from './ui/EditProfileForm/EditProfileForm'
+export { ProfilePhoto } from './ui/ProfilePhoto/ProfilePhoto'
