--- src/feature/profile/ui/AddProfilePhotoDialog/AddProfilePhotoDialog.module.scss
@@ -1,5 +1,0 @@
-.dialog {
-  width: 100%;
-  max-width: 492px;
-  margin: 0 15px;
-}
