+++ src/feature/profile/ui/ProfilePhoto/ProfilePhoto.module.scss
@@ -62,3 +62,9 @@
   width: 16px;
   height: 16px;
 }
+
+.dialog {
+  width: 100%;
+  max-width: 492px;
+  margin: 0 15px;
+}
