+++ src/pages/profile/[id]/settings/index.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useMemo, useState } from 'react'
 
-import { GeneralInformation } from '@/feature/profile'
 import { DefenderProtectedRoute } from '@/shared/helpers'
 import { useTranslation } from '@/shared/hooks'
 import { Page } from '@/shared/types/layout'
 import { Tabs } from '@/shared/ui/Tabs'
 import { SidebarLayout } from '@/widgets/layout'
+import { GeneralInformation } from '@/widgets/profile'
 import { useRouter } from 'next/router'
 
 const valuesTabs = ['general', 'devices', 'management', 'payments']
