+++ src/widgets/post/index.ts
@@ -1,0 +1,2 @@
+export { CreatePostDialog } from './ui/CreatePostDialog/CreatePostDialog'
+export { PostDetailsDialogs } from './ui/PostDetailsDialogs/PostDetailsDialogs'
