+++ src/widgets/post/ui/CreatePostDialog/CreatePostDialog.module.scss
