+++ src/widgets/post/ui/CreatePostDialog/CreatePostDialog.tsx
@@ -3,6 +3,7 @@ import { toast } from 'react-toastify'
 
 import { ConfirmDialog } from '@/entities/dialog'
 import { MAX_SIZE_IMAGE_20MB, postsActions } from '@/entities/post'
+import { CropperPostScreen, CurrentWindow, DescriptionScreen, FiltersScreen } from '@/feature/post'
 import { getModifiedImage, photoSchema } from '@/shared/helpers'
 import { useAppDispatch, useAppSelector, useTranslation } from '@/shared/hooks'
 import { Button } from '@/shared/ui/Button'
@@ -12,10 +13,6 @@ import clsx from 'clsx'
 
 import s from './CreatePostDialog.module.scss'
 
-import { CurrentWindow } from '../../model/types/post.types'
-import { CropperPostScreen } from '../CropperPostScreen/CropperPostScreen'
-import { DescriptionScreen } from '../DescriptionScreen/DescriptionScreen'
-import { FiltersScreen } from '../FiltersScreen/FiltersScreen'
 import { Title } from './Title/Title'
 
 type Props = {
