+++ src/widgets/post/ui/CreatePostDialog/Title/Title.module.scss
