+++ src/widgets/post/ui/PostDetailsDialogs/PostDetailsDialogs.module.scss
