+++ src/widgets/profile/index.ts
@@ -1,2 +1,3 @@
+export { GeneralInformation } from './ui/GeneralInformation/GeneralInformation'
 export { MainProfileContent } from './ui/MainProfileContent/MainProfileContent'
 export { SomeProfileContent } from './ui/SomeProfileContent/SomeProfileContent'
