+++ src/widgets/profile/ui/GeneralInformation/GeneralInformation.module.scss
